Refactor notification packages
