+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/notification/DailyStudyReminderNotificationDelegate.kt
@@ -4,15 +4,15 @@ import android.content.Context
 import androidx.core.app.NotificationCompat
 import androidx.core.content.ContextCompat
 import java.text.SimpleDateFormat
-import java.util.Calendar
+import java.util.*
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.analytic.domain.model.hyperskill.HyperskillAnalyticRoute
 import org.hyperskill.app.android.R
 import org.hyperskill.app.android.core.extensions.DateTimeHelper
 import org.hyperskill.app.android.notification.model.DailyStudyReminderClickedData
 import org.hyperskill.app.android.notification.model.HyperskillNotificationChannel
-import org.hyperskill.app.notification.domain.analytic.NotificationDailyStudyReminderShownHyperskillAnalyticEvent
-import org.hyperskill.app.notification.domain.interactor.NotificationInteractor
+import org.hyperskill.app.notification.local.domain.analytic.NotificationDailyStudyReminderShownHyperskillAnalyticEvent
+import org.hyperskill.app.notification.local.domain.interactor.NotificationInteractor
 
 class DailyStudyReminderNotificationDelegate(
     hyperskillNotificationManager: HyperskillNotificationManager,
