+++ shared/src/commonMain/kotlin/org/hyperskill/app/main/domain/interactor/AppInteractor.kt
@@ -2,11 +2,11 @@ package org.hyperskill.app.main.domain.interactor
 
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.auth.domain.interactor.AuthInteractor
+import org.hyperskill.app.notification.remote.domain.interactor.PushNotificationsInteractor
 import org.hyperskill.app.profile.domain.interactor.ProfileInteractor
 import org.hyperskill.app.progresses.domain.repository.ProgressesRepository
 import org.hyperskill.app.projects.domain.repository.ProjectsRepository
 import org.hyperskill.app.providers.domain.repository.ProvidersRepository
-import org.hyperskill.app.push_notifications.domain.interactor.PushNotificationsInteractor
 import org.hyperskill.app.track.domain.repository.TrackRepository
 import org.hyperskill.app.user_storage.domain.interactor.UserStorageInteractor
