--- shared/src/commonMain/kotlin/org/hyperskill/app/notification/data/model/NotificationDescription.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.notification.data.model
-
-data class NotificationDescription(
-    val id: Int,
-    val title: String,
-    val text: String
-)
