--- shared/src/commonMain/kotlin/org/hyperskill/app/notification/domain/flow/DailyStudyRemindersEnabledFlow.kt
@@ -1,5 +1,0 @@
-package org.hyperskill.app.notification.domain.flow
-
-import org.hyperskill.app.core.domain.flow.SharedDataFlow
-
-interface DailyStudyRemindersEnabledFlow : SharedDataFlow<Boolean>
