--- shared/src/commonMain/kotlin/org/hyperskill/app/notification/injection/NotificationComponent.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.notification.injection
-
-import org.hyperskill.app.notification.domain.interactor.NotificationInteractor
-
-interface NotificationComponent {
-    val notificationInteractor: NotificationInteractor
-}
