--- shared/src/commonMain/kotlin/org/hyperskill/app/notification/injection/NotificationFlowDataComponent.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.notification.injection
-
-import org.hyperskill.app.notification.domain.flow.DailyStudyRemindersEnabledFlow
-
-interface NotificationFlowDataComponent {
-    val dailyStudyRemindersEnabledFlow: DailyStudyRemindersEnabledFlow
-}
