--- shared/src/commonMain/kotlin/org/hyperskill/app/notification/injection/NotificationFlowDataComponentImpl.kt
@@ -1,9 +1,0 @@
-package org.hyperskill.app.notification.injection
-
-import org.hyperskill.app.notification.data.flow.DailyStudyRemindersEnabledFlowImpl
-import org.hyperskill.app.notification.domain.flow.DailyStudyRemindersEnabledFlow
-
-class NotificationFlowDataComponentImpl : NotificationFlowDataComponent {
-    override val dailyStudyRemindersEnabledFlow: DailyStudyRemindersEnabledFlow =
-        DailyStudyRemindersEnabledFlowImpl()
-}
