+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/local/data/model/NotificationDescription.kt
@@ -1,0 +1,7 @@
+package org.hyperskill.app.notification.local.data.model
+
+data class NotificationDescription(
+    val id: Int,
+    val title: String,
+    val text: String
+)
