+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/local/domain/flow/DailyStudyRemindersEnabledFlow.kt
@@ -1,0 +1,5 @@
+package org.hyperskill.app.notification.local.domain.flow
+
+import org.hyperskill.app.core.domain.flow.SharedDataFlow
+
+interface DailyStudyRemindersEnabledFlow : SharedDataFlow<Boolean>
