+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/local/injection/NotificationComponent.kt
@@ -1,0 +1,7 @@
+package org.hyperskill.app.notification.local.injection
+
+import org.hyperskill.app.notification.local.domain.interactor.NotificationInteractor
+
+interface NotificationComponent {
+    val notificationInteractor: NotificationInteractor
+}
