+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/local/injection/NotificationComponentImpl.kt
@@ -1,0 +1,23 @@
+package org.hyperskill.app.notification.local.injection
+
+import org.hyperskill.app.core.injection.AppGraph
+import org.hyperskill.app.notification.local.cache.NotificationCacheDataSourceImpl
+import org.hyperskill.app.notification.local.data.repository.NotificationRepositoryImpl
+import org.hyperskill.app.notification.local.data.source.NotificationCacheDataSource
+import org.hyperskill.app.notification.local.domain.interactor.NotificationInteractor
+import org.hyperskill.app.notification.local.domain.repository.NotificationRepository
+
+class NotificationComponentImpl(appGraph: AppGraph) : NotificationComponent {
+    private val notificationCacheDataSource: NotificationCacheDataSource =
+        NotificationCacheDataSourceImpl(appGraph.commonComponent.settings, appGraph.commonComponent.resourceProvider)
+
+    private val notificationRepository: NotificationRepository =
+        NotificationRepositoryImpl(notificationCacheDataSource)
+
+    override val notificationInteractor: NotificationInteractor =
+        NotificationInteractor(
+            notificationRepository,
+            appGraph.submissionDataComponent.submissionRepository,
+            appGraph.notificationFlowDataComponent.dailyStudyRemindersEnabledFlow
+        )
+}
