+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/local/injection/NotificationFlowDataComponent.kt
@@ -1,0 +1,7 @@
+package org.hyperskill.app.notification.local.injection
+
+import org.hyperskill.app.notification.local.domain.flow.DailyStudyRemindersEnabledFlow
+
+interface NotificationFlowDataComponent {
+    val dailyStudyRemindersEnabledFlow: DailyStudyRemindersEnabledFlow
+}
