+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/local/injection/NotificationFlowDataComponentImpl.kt
@@ -1,0 +1,9 @@
+package org.hyperskill.app.notification.local.injection
+
+import org.hyperskill.app.notification.local.data.flow.DailyStudyRemindersEnabledFlowImpl
+import org.hyperskill.app.notification.local.domain.flow.DailyStudyRemindersEnabledFlow
+
+class NotificationFlowDataComponentImpl : NotificationFlowDataComponent {
+    override val dailyStudyRemindersEnabledFlow: DailyStudyRemindersEnabledFlow =
+        DailyStudyRemindersEnabledFlowImpl()
+}
