+++ shared/src/commonMain/kotlin/org/hyperskill/app/notification/remote/injection/PushNotificationsComponent.kt
@@ -1,0 +1,7 @@
+package org.hyperskill.app.notification.remote.injection
+
+import org.hyperskill.app.notification.remote.domain.interactor.PushNotificationsInteractor
+
+interface PushNotificationsComponent {
+    val pushNotificationsInteractor: PushNotificationsInteractor
+}
