+++ shared/src/commonMain/kotlin/org/hyperskill/app/profile/injection/ProfileFeatureBuilder.kt
@@ -3,8 +3,8 @@ package org.hyperskill.app.profile.injection
 import org.hyperskill.app.analytic.domain.interactor.AnalyticInteractor
 import org.hyperskill.app.core.presentation.ActionDispatcherOptions
 import org.hyperskill.app.magic_links.domain.interactor.UrlPathProcessor
-import org.hyperskill.app.notification.domain.flow.DailyStudyRemindersEnabledFlow
-import org.hyperskill.app.notification.domain.interactor.NotificationInteractor
+import org.hyperskill.app.notification.local.domain.flow.DailyStudyRemindersEnabledFlow
+import org.hyperskill.app.notification.local.domain.interactor.NotificationInteractor
 import org.hyperskill.app.products.domain.interactor.ProductsInteractor
 import org.hyperskill.app.profile.domain.interactor.ProfileInteractor
 import org.hyperskill.app.profile.presentation.ProfileActionDispatcher
