--- shared/src/commonMain/kotlin/org/hyperskill/app/push_notifications/injection/PushNotificationsComponent.kt
@@ -1,7 +1,0 @@
-package org.hyperskill.app.push_notifications.injection
-
-import org.hyperskill.app.push_notifications.domain.interactor.PushNotificationsInteractor
-
-interface PushNotificationsComponent {
-    val pushNotificationsInteractor: PushNotificationsInteractor
-}
